+++ src/Cpu.py
@@ -58,6 +58,9 @@ class Cpu:
         # Set event masks for drawingarea in order to enable these events.
         self.drawingarea1101.set_events(Gdk.EventMask.ENTER_NOTIFY_MASK | Gdk.EventMask.LEAVE_NOTIFY_MASK | Gdk.EventMask.POINTER_MOTION_MASK)
 
+        # "0" value of "initial_already_run" variable means that initial function is not run before or tab settings are reset from general settings and initial function have to be run.
+        self.initial_already_run = 0
+
 
     # ----------------------- "customizations menu" Button -----------------------
     def on_button1101_clicked(self, widget):
@@ -117,6 +120,17 @@ class Cpu:
         self.selected_cpu_core_prev = selected_cpu_core
 
         self.drawingarea1101.queue_draw()
+
+        # Run "main_gui_device_selection_list_func" if selected device list is changed since the last loop.
+        logical_core_list_system_ordered = Performance.logical_core_list_system_ordered
+        try:                                                                                      
+            if self.logical_core_list_system_ordered_prev != logical_core_list_system_ordered:
+                from MainGUI import MainGUI
+                MainGUI.main_gui_device_selection_list_func()
+        # try-except is used in order to avoid error and also run "main_gui_device_selection_list_func" if this is first loop of the function.
+        except AttributeError:
+            pass
+        self.logical_core_list_system_ordered_prev = logical_core_list_system_ordered
 
 
         # Get information.